initial definition of Surface trait
